requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
transformers
torch
accelerate
protobuf==3.20.3
